First running version
There is a problem in the example model:

When both color and textColor are undefined (and there is some text so
that textColor is active), the default computation for each of these
cstics depends on the value of the other.  This causes a circular
dependency.

Furthermore, if, say, color is explicitly "red" and textColor is "green"
by default, then the choice color="green" is marked as erroneous, even
though choosing it would not cause an error since then the default of
textColor would switch to "red".